Drop the dependency on the deprecated CryptoHash
The [cryptohash package](https://hackage.haskell.org/package/cryptohash) has been deprecated and replaced with cryptonite, which this package is already using. Thus, it makes sense to drop the dependency on cryptohash and just use cryptonite.

In cryptonite, the hashing functions return a `Digest` instead of a `ByteString`. Both are instances of `ByteArray` (from the memory package), so we can convert between them using Data.ByteArray.convert. memory is already a dependency via cryptonite, so no new dependency is added.

memory doesn't make any API guarantees, so I'm not sure what's best to put as a version bound. It could be constrained to just the latest version as of now (0.14.16), but that will necessitate updating the u2f package whenever a new memory package comes out. That's fine, it just might be more maitenance hassle than it's worth. I'm open to whatever.